Adds Per Provider Seed Ratio
Adds Per Provider Seed Ratio for all torrent providers apart from users
custom rss feeds (I will add this at a later date).
Support for utorrent, transmission and deluge is included.

Some minor bug fixes are also included:
Fixes deluge hang when adding a label with spaces

The deluge client cannot handle labels with spaces. It currently causes
a number of timeout errors if you try to add a label with a space.
This commit adds a code to reject labels with spaces in and also adds a
warning box in the deluge torrent page warning users not to create
labels with spaces in them i.e. "tv shows" is not allowed but "tvshows"
is fine.

Since rtorrent seed ratio has been disabled in sickbeard for a long time
(the code is hashed out in the master branch) I have hidden the ratio
setting box in the rtorrent config page until it gets fixed.
There is apparently a bug in rtorrent that causes random torrents to enter
a pause state when setting the seed ratio.
I have also added a method to each torrent client to capture converting
to float errors when setting the seed ratio.