Fixes deluge hang when adding a label with spaces
The deluge client cannot handle labels with spaces. It currently causes
a number of timeout errors if you try to add a label with a space.
This commit adds a code to reject labels with spaces in and also adds a
warning box in the deluge torrent page warning users not to create
labels with spaces in them i.e. "tv shows" is not allowed but "tvshows"
is fine.