Fixed issue where consecutive manual searches where not queued properly Added first backend processing for retrieving all queued searches. This should fix the frondend blocking when doing manual searches for episodes, because the backend is free sooner. It now only queues the search. TODO: Create a returning ajax call for getting a list of all searches in queue en running on the displayShow page.
Fixed issue where consecutive manual searches where not queued properly
Added first backend processing for retrieving all queued searches. This should fix the frondend blocking when doing manual searches for episodes, because the backend is free sooner. It now only queues the search.
TODO: Create a returning ajax call for getting a list of all searches in queue en running on the displayShow page.

Implemented first version of frontend for querying the Manual search status.
It's still buggy, as I don't have a way of knowing if a search for a show in the backend has finished, and what the status was of the search.

Added first functionality for keeping track of also the finished searches.
Need to figure out how to keep priority higher for running and queued searches.

Some fixes in frontend en backend for making sure finished downloads show status/quality again.
Retry functionality is broken though.

Got the alert prompt for marking a download as failed and retry back. But really don't know what the use is of this? It just only prompts.. no different functionality in backend, compared to a normal search.

Small fixes to the ui, for updating the searches, queuing animations.

Soms code cleanup