Travis-CI has not actually been running the tests.
When ./all_tests.py is run from outside of the tests directory it fails to run the
majority of the test suite. This is either something that needs to be fixed in
all_tests.py or we need to cd into the tests dir before running them.

Also introduce a test blacklist which will make it easier to avoid tests that you
do not want to have run on a regular basis.