Fix Plex notification status check
The return value of update_library() was changed from True/False to None/Error string in https://github.com/SiCKRAGETV/SickRage/commit/4c9d9185b0c25d435a6bbb20efa60b1f4337a016 but the check for the return value was never updated. This corrects that so the notification sent out correctly states that it sent a notification to the server instead of defaulting to stating it was unable to connect.