Exit function properly on exception
When "metadata.extract(parser)" raises an exception, the function exits without returning "None", which misleads the calling function that it exited with usably data, and it then interprets the result incorrectly.

The end result is that image_cache.py > fill_cache(self, show_obj) exits early on an exception when attempting to retrieve images for the cache from a meta data provider, instead of continuing to the second step of attempting to retrieve images for the cache from the indexer online.  

Without this fix the log gets the generic error message "bad descriptor" for an unknown error from "metadata.extract(parser)", and the secondary cache image retrieval attempt from the indexer is never made.