Fix SNATCHED_PROPER and SNATCHED_BEST not changing status to DOWNLOADED
At the first scan, the file and file size are detected. As snatched_proper and snatched_best  status wasn't checke, file size were updated in DB and in the next scan line 721 ("elif not same_file ... ), the if is false as SR detects the same_size = True so ep status stayed forever as SNATCHED_PROPER/BEST

This is the real fix for SiCKRAGETV/sickrage-issues/issues/1074

Also changed some log messages for better understanding as "record is not dirty" it's not clear for average user.

Working without issues for this show that only downloaded proper and never changed status to downloaded.
```
2015-10-19 10:53:24 INFO     SHOWQUEUE-REFRESH :: Show Les Revenants S02E06 used to be Snatched (Proper) (720p HDTV) but a file exists with quality 720p HDTV so I'm setting the status to Downloaded (720p HDTV)
2015-10-19 10:53:24 INFO     SHOWQUEUE-REFRESH :: Show Les Revenants S02E05 used to be Snatched (Proper) (720p HDTV) but a file exists with quality 720p HDTV so I'm setting the status to Downloaded (720p HDTV)
2015-10-19 10:53:23 INFO     SHOWQUEUE-REFRESH :: Performing refresh on Les Revenants
```

To reproduce the issue, just add this show and enable provider T411
@duramato can you test/confirm, please?