Do not make torrent labels lower case.
This can cause some issues when using rTorrent, as the upper and lower case labels are seen as two different labels, and will tell the client to move files into the wrong directory if the file system is case-sensitive. If the user wishes to keep the label lower-case, they can input a lower case label name into the field. However, the label should be accepted as-is.