Fix API returning error on help
Attempting to pass `args` and `kwargs` without the preceeding `*` && `**` is broken. If you don't have it then they are just additional arguments that get passed to the special `*args` variable of the called function.  This was causing `kwargs` to get mangled into `args` and then leading to an error about how it couldn't parse the next param.  You can see this by doing `/api/[yourkey]/?cmd=show.addnew&help=1`.

I am sure this was causing other errors too, but help was the first one I saw :)